Create VM test
